+++ src/emhass/utils.py
@@ -136,6 +136,7 @@ def treat_runtimeparams(runtimeparams: str, params: str, retrieve_hass_conf: dic
                            'custom_batt_soc_forecast_id': {"entity_id": "sensor.soc_batt_forecast", "unit_of_measurement": "%", "friendly_name": "Battery SOC Forecast"},
                            'custom_grid_forecast_id': {"entity_id": "sensor.p_grid_forecast", "unit_of_measurement": "W", "friendly_name": "Grid Power Forecast"},
                            'custom_cost_fun_id': {"entity_id": "sensor.total_cost_fun_value", "unit_of_measurement": "", "friendly_name": "Total cost function value"},
+                           'custom_optim_status_id': {"entity_id": "sensor.optim_status", "unit_of_measurement": "", "friendly_name": "EMHASS optimization status"},
                            'custom_unit_load_cost_id': {"entity_id": "sensor.unit_load_cost", "unit_of_measurement": "€/kWh", "friendly_name": "Unit Load Cost"},
                            'custom_unit_prod_price_id': {"entity_id": "sensor.unit_prod_price", "unit_of_measurement": "€/kWh", "friendly_name": "Unit Prod Price"},
                            'custom_deferrable_forecast_id': custom_deferrable_forecast_id,
@@ -339,6 +340,8 @@ def treat_runtimeparams(runtimeparams: str, params: str, retrieve_hass_conf: dic
             params['passed_data']['custom_grid_forecast_id'] = runtimeparams['custom_grid_forecast_id']
         if 'custom_cost_fun_id' in runtimeparams.keys():
             params['passed_data']['custom_cost_fun_id'] = runtimeparams['custom_cost_fun_id']
+        if 'custom_optim_status_id' in runtimeparams.keys():
+            params['passed_data']['custom_optim_status_id'] = runtimeparams['custom_optim_status_id']
         if 'custom_unit_load_cost_id' in runtimeparams.keys():
             params['passed_data']['custom_unit_load_cost_id'] = runtimeparams['custom_unit_load_cost_id']
         if 'custom_unit_prod_price_id' in runtimeparams.keys():